+++ @planetarium/tx/src/tx/unsigned.ts
@@ -1,11 +1,16 @@
 import { Encodable } from "bencodex";
-import { encodeSystemAction, SystemAction } from "../action";
+import { CustomAction, encodeSystemAction, SystemAction } from "../action";
 import { encodeTxMetadata, TxMetadata } from "./metadata";
 
 const SYSTEM_ACTION_KEY = Buffer.from([0x41]); // 'A'
+const CUSTOM_ACTION_KEY = Buffer.from([0x61]); // 'a'
 
 export interface UnsignedTxWithSystemAction extends TxMetadata {
   systemAction: SystemAction;
+}
+
+export interface UnsignedTxWithCustomActions extends TxMetadata {
+  customActions: CustomAction[];
 }
 
 /**
@@ -18,5 +23,18 @@ export function encodeUnsignedTxWithSystemAction(
 ): Encodable {
   const dict = encodeTxMetadata(metadata);
   dict.set(SYSTEM_ACTION_KEY, encodeSystemAction(metadata.systemAction));
+  return dict;
+}
+
+/**
+ * Encodes an unsigned transaction with custom actions.
+ * @param tx An unsigned transaction with custom actions.
+ * @returns An encoded transaction.
+ */
+export function encodeUnsignedTxWithCustomActions(
+  metadata: UnsignedTxWithCustomActions
+): Encodable {
+  const dict = encodeTxMetadata(metadata);
+  dict.set(CUSTOM_ACTION_KEY, metadata.customActions);
   return dict;
 }