+++ Libplanet.Net/Transports/InvalidMessageSignatureException.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Runtime.Serialization;
-using Bencodex;
 using Libplanet.Crypto;
 using Libplanet.Net.Messages;
 using Libplanet.Serialization;
@@ -14,8 +13,6 @@ namespace Libplanet.Net.Transports
     [Serializable]
     public class InvalidMessageSignatureException : Exception
     {
-        private static Codec _codec = new Codec();
-
         internal InvalidMessageSignatureException(
             string message,
             BoundPeer peer,
@@ -35,7 +32,7 @@ namespace Libplanet.Net.Transports
             StreamingContext context)
             : base(info, context)
         {
-            Peer = new BoundPeer(_codec.Decode(info.GetValue<byte[]>(nameof(Peer))));
+            Peer = info.GetValue<BoundPeer>(nameof(Peer));
             PublicKey = new PublicKey(info.GetValue<byte[]>(nameof(PublicKey)));
             MessageToVerify = info.GetValue<byte[]>(nameof(MessageToVerify));
             Signature = info.GetValue<byte[]>(nameof(Signature));
@@ -53,7 +50,7 @@ namespace Libplanet.Net.Transports
             SerializationInfo info, StreamingContext context)
         {
             base.GetObjectData(info, context);
-            info.AddValue(nameof(Peer), _codec.Encode(Peer.Bencoded));
+            info.AddValue(nameof(Peer), Peer);
             info.AddValue(nameof(PublicKey), PublicKey.Format(true));
             info.AddValue(nameof(MessageToVerify), MessageToVerify);
             info.AddValue(nameof(Signature), Signature);