+++ Libplanet/Address.cs
@@ -4,6 +4,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
 using System.Globalization;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -40,9 +41,10 @@ namespace Libplanet
     /// </summary>
     /// <remarks>Every <see cref="Address"/> value is immutable.</remarks>
     /// <seealso cref="PublicKey"/>
+    [Serializable]
     [JsonConverter(typeof(AddressJsonConverter))]
     public readonly struct Address
-        : IEquatable<Address>, IComparable<Address>, IComparable, IBencodable
+        : ISerializable, IEquatable<Address>, IComparable<Address>, IComparable, IBencodable
     {
         /// <summary>
         /// The <see cref="byte"/>s size that each <see cref="Address"/> takes.
@@ -50,7 +52,10 @@ namespace Libplanet
         /// </summary>
         public const int Size = 20;
 
-        private static readonly byte[] _defaultByteArray = new byte[Size];
+        private static readonly Codec _codec = new Codec();
+
+        private static readonly ImmutableArray<byte> _defaultByteArray =
+            ImmutableArray.Create<byte>(new byte[Size]);
 
         private readonly ImmutableArray<byte> _byteArray;
 
@@ -70,7 +75,8 @@ namespace Libplanet
         {
             if (address.Length != Size)
             {
-                throw new ArgumentException("address must be 20 bytes", nameof(address));
+                throw new ArgumentException(
+                    $"Given {nameof(address)} must be 20 bytes", nameof(address));
             }
 
             _byteArray = address;
@@ -81,10 +87,7 @@ namespace Libplanet
         /// cref="byte"/> array (i.e., <paramref name="address"/>).
         /// </summary>
         /// <param name="address">An array of 20 <see cref="byte"/>s which
-        /// represents an <see cref="Address"/>.  This must not be <see langword="null"/>.
-        /// </param>
-        /// <exception cref="ArgumentNullException">Thrown when <see langword="null"/> was
-        /// passed to <paramref name="address"/>.</exception>
+        /// represents an <see cref="Address"/>.</param>
         /// <exception cref="ArgumentException">Thrown when the given <paramref
         /// name="address"/> array did not lengthen 20 bytes.</exception>
         /// <remarks>A valid <see cref="byte"/> array which represents an
@@ -92,7 +95,7 @@ namespace Libplanet
         /// /> method.</remarks>
         /// <seealso cref="ToByteArray()"/>
         public Address(byte[] address)
-            : this(address?.ToImmutableArray() ?? throw new ArgumentNullException(nameof(address)))
+            : this(address.ToImmutableArray())
         {
         }
 
@@ -140,18 +143,28 @@ namespace Libplanet
         /// </param>
         /// <exception cref="ArgumentException">Thrown when given <paramref name="bencoded"/>
         /// is not of type <see cref="Binary"/>.</exception>
+        /// <seealso cref="Address(in ImmutableArray{byte})"/>
         public Address(IValue bencoded)
             : this(bencoded is Binary binary
-                ? (Binary)bencoded
+                ? binary
                 : throw new ArgumentException(
                     $"Given {nameof(bencoded)} must be of type " +
-                    $"{typeof(Bencodex.Types.Binary)}: {bencoded.GetType()}",
+                    $"{typeof(Binary)}: {bencoded.GetType()}",
                     nameof(bencoded)))
         {
         }
 
         private Address(Binary bencoded)
             : this(bencoded.ByteArray)
+        {
+        }
+
+        private Address(SerializationInfo info, StreamingContext context)
+            : this(_codec.Decode(info.GetValue(nameof(Bencoded), typeof(byte[])) is
+                byte[] bytes
+                    ? bytes
+                    : throw new SerializationException(
+                        $"Invalid type for {nameof(Bencoded)} in {nameof(info)}.")))
         {
         }
 
@@ -162,10 +175,9 @@ namespace Libplanet
         /// <remarks>This is immutable.  For a mutable array, call <see
         /// cref="ToByteArray()"/> method.</remarks>
         /// <seealso cref="ToByteArray()"/>
-        public ImmutableArray<byte> ByteArray =>
-            _byteArray.IsDefault
-                ? _defaultByteArray.ToImmutableArray()
-                : _byteArray;
+        public ImmutableArray<byte> ByteArray => _byteArray.IsDefault
+            ? _defaultByteArray
+            : _byteArray;
 
         /// <inheritdoc/>
         public IValue Bencoded => new Binary(ByteArray);
@@ -191,9 +203,7 @@ namespace Libplanet
         /// <seealso cref="ByteArray"/>
         /// <seealso cref="Address(byte[])"/>
         [Pure]
-        public byte[] ToByteArray() => ByteArray.IsDefault
-            ? _defaultByteArray
-            : ByteArray.ToArray();
+        public byte[] ToByteArray() => ByteArray.ToArray();
 
         /// <summary>
         /// Gets a mixed-case hexadecimal string of 40 letters that represent
@@ -212,11 +222,7 @@ namespace Libplanet
         /// method instead.</remarks>
         /// <seealso cref="ToString()"/>
         [Pure]
-        public string ToHex()
-        {
-            string hex = ByteUtil.Hex(ToByteArray());
-            return ToChecksumAddress(hex);
-        }
+        public string ToHex() => ToChecksumAddress(ByteUtil.Hex(ToByteArray()));
 
         /// <summary>
         /// Gets a <c>0x</c>-prefixed mixed-case hexadecimal string of
@@ -235,9 +241,12 @@ namespace Libplanet
         /// instead.</remarks>
         /// <seealso cref="ToHex()"/>
         [Pure]
-        public override string ToString()
+        public override string ToString() => $"0x{ToHex()}";
+
+        /// <inheritdoc/>
+        public void GetObjectData(SerializationInfo info, StreamingContext context)
         {
-            return $"0x{ToHex()}";
+            info.AddValue(nameof(Bencoded), _codec.Encode(Bencoded));
         }
 
         /// <inheritdoc cref="IComparable{T}.CompareTo(T)"/>
@@ -351,10 +360,10 @@ namespace Libplanet
             JsonSerializerOptions options
         )
         {
-            string? hex = reader.GetString();
+            string hex = reader.GetString() ?? throw new JsonException("Expected a string.");
             try
             {
-                return new Address(hex!);
+                return new Address(hex);
             }
             catch (ArgumentException e)
             {